backend/requirements.txt
fastapi
uvicorn
openai-whisper
torch